split into matrix.h + matrix_stack.h
