+++ src/code/z_skin_matrix.c
@@ -1,5 +1,6 @@
 #include "global.h"
 #include "terminal.h"
+#include "matrix.h"
 
 // clang-format off
 MtxF sMtxFClear = {
